Revert "feat(inertia): introduce CLI commands for build, dev, install, and types"
This reverts commit 093eb0021853cab11c4b9e83500e66f2e113d239.

The swap from Quarry inertia:* commands to a standalone `inertia` bin
was made to work around an OOM, but the OOM still reproduces with the
standalone bin. Quarry was not the cause, so restore the original
command-based layout while we continue investigating.

Also drops the now-stale CLI-bin bullet from packages/inertia/CLAUDE.md.
