+++ packages/inertia/src/inertia.module.ts
@@ -5,6 +5,10 @@ import type { AsyncModuleOptions, DynamicModule, OnException, OnInitialize } fro
 import { Module } from 'stratal/module'
 import { SchemaValidationError, type RouteConfigurable, type Router } from 'stratal/router'
 import { augmentRouterContext } from './augment/router-context'
+import { InertiaBuildCommand } from './commands/inertia-build.command'
+import { InertiaDevCommand } from './commands/inertia-dev.command'
+import { InertiaInstallCommand } from './commands/inertia-install.command'
+import { InertiaTypesCommand } from './commands/inertia-types.command'
 import type { InertiaModuleOptions } from './inertia.options'
 import { INERTIA_TOKENS } from './inertia.tokens'
 import { InertiaMiddleware } from './middleware/inertia.middleware'
@@ -19,6 +23,10 @@ import { TemplateService } from './services/template.service'
     { provide: INERTIA_TOKENS.TemplateService, useClass: TemplateService },
     { provide: INERTIA_TOKENS.ManifestService, useClass: ManifestService },
     { provide: INERTIA_TOKENS.SsrRenderer, useClass: SsrRendererService, scope: Scope.Singleton },
+    InertiaInstallCommand,
+    InertiaTypesCommand,
+    InertiaDevCommand,
+    InertiaBuildCommand,
   ],
 })
 export class InertiaModule implements RouteConfigurable, OnInitialize, OnException {
